requirements_win.txt
multipledispatch==1.0.0
numpy==1.26.4
pillow==10.2.0
PyOpenGL==3.1.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrr==0.10.3
six==1.16.0
